+++ packages/expo-updates/android/src/androidTest/java/expo/modules/updates/manifest/LegacyManifestTest.java
@@ -32,7 +32,7 @@ public class LegacyManifestTest {
   }
 
   @Test
-  public void testGetAssetsUrlBase_assetUrlOverride_absoluteUrl() throws JSONException {
+  public void testGetAssetsUrlBase_AssetUrlOverride_AbsoluteUrl() throws JSONException {
     String assetUrlBase = "https://xxx.dev/~assets";
 
     Uri manifestUrl = Uri.parse("https://esamelson.github.io/self-hosting-test/android-index.json");
@@ -45,7 +45,7 @@ public class LegacyManifestTest {
   }
 
   @Test
-  public void testGetAssetsUrlBase_assetUrlOverride_relativeUrl() throws JSONException {
+  public void testGetAssetsUrlBase_AssetUrlOverride_RelativeUrl() throws JSONException {
     Uri manifestUrl = Uri.parse("https://esamelson.github.io/self-hosting-test/android-index.json");
     JSONObject manifestJson = new JSONObject();
     manifestJson.put("assetUrlOverride", "my_assets");
@@ -56,7 +56,18 @@ public class LegacyManifestTest {
   }
 
   @Test
-  public void testGetAssetsUrlBase_assetUrlOverride_relativeUrlDotSlash() throws JSONException {
+  public void testGetAssetsUrlBase_AssetUrlOverride_OriginRelativeUrl() throws JSONException {
+    Uri manifestUrl = Uri.parse("https://esamelson.github.io/self-hosting-test/android-index.json");
+    JSONObject manifestJson = new JSONObject();
+    manifestJson.put("assetUrlOverride", "/my_assets");
+
+    Uri expected = Uri.parse("https://esamelson.github.io/my_assets");
+    Uri actual = LegacyManifest.getAssetsUrlBase(manifestUrl, manifestJson);
+    Assert.assertEquals(expected, actual);
+  }
+
+  @Test
+  public void testGetAssetsUrlBase_AssetUrlOverride_RelativeUrlDotSlash() throws JSONException {
     Uri manifestUrl = Uri.parse("https://esamelson.github.io/self-hosting-test/android-index.json");
     JSONObject manifestJson = new JSONObject();
     manifestJson.put("assetUrlOverride", "./my_assets");
@@ -67,7 +78,40 @@ public class LegacyManifestTest {
   }
 
   @Test
-  public void testGetAssetsUrlBase_assetUrlOverride_default() {
+  public void testGetAssetsUrlBase_AssetUrlOverride_Normalize() throws JSONException {
+    Uri manifestUrl = Uri.parse("https://esamelson.github.io/self-hosting-test/android-index.json");
+    JSONObject manifestJson = new JSONObject();
+    manifestJson.put("assetUrlOverride", "./a/../b");
+
+    Uri expected = Uri.parse("https://esamelson.github.io/self-hosting-test/b");
+    Uri actual = LegacyManifest.getAssetsUrlBase(manifestUrl, manifestJson);
+    Assert.assertEquals(expected, actual);
+  }
+
+  @Test
+  public void testGetAssetsUrlBase_AssetUrlOverride_NormalizeToHostname() throws JSONException {
+    Uri manifestUrl = Uri.parse("https://esamelson.github.io/self-hosting-test/android-index.json");
+    JSONObject manifestJson = new JSONObject();
+    manifestJson.put("assetUrlOverride", "../b");
+
+    Uri expected = Uri.parse("https://esamelson.github.io/b");
+    Uri actual = LegacyManifest.getAssetsUrlBase(manifestUrl, manifestJson);
+    Assert.assertEquals(expected, actual);
+  }
+
+  @Test
+  public void testGetAssetsUrlBase_AssetUrlOverride_NormalizePastHostname() throws JSONException {
+    Uri manifestUrl = Uri.parse("https://esamelson.github.io/self-hosting-test/android-index.json");
+    JSONObject manifestJson = new JSONObject();
+    manifestJson.put("assetUrlOverride", "../../b");
+
+    Uri expected = Uri.parse("https://esamelson.github.io/b");
+    Uri actual = LegacyManifest.getAssetsUrlBase(manifestUrl, manifestJson);
+    Assert.assertEquals(expected, actual);
+  }
+
+  @Test
+  public void testGetAssetsUrlBase_AssetUrlOverride_Default() {
     Uri manifestUrl = Uri.parse("https://esamelson.github.io/self-hosting-test/android-index.json");
     JSONObject manifestJson = new JSONObject();